Untangle paftl from pafmath and p2dpoly
Required splitting off LinReg (linear regression) from pafmath and BSPNode from p2dpoly
