+++ genlib/pafmath.cpp
@@ -20,15 +20,16 @@
 
 #include <math.h>
 #include <genlib/pafmath.h>
+#include <inttypes.h>
 
-uint64 g_rand[11] = {1,2,3,5,7,11,13,17,19,23,29};
+uint64_t g_rand[11] = {1,2,3,5,7,11,13,17,19,23,29};
 
 // 25-Jul-2007: changed the g_mult and g_const used for random number generation
 // for some reason, there appeared to be a pattern to the numbers
 
 // Quick mod - TV
-const uint64 g_mult  = /*(0xF9561B2E << 32) + */0x71A7FA85;
-const uint64 g_const = /*(0x9BB3920E << 32) + */0xF5E958B9;
+const uint64_t g_mult  = /*(0xF9561B2E << 32) + */0x71A7FA85;
+const uint64_t g_const = /*(0x9BB3920E << 32) + */0xF5E958B9;
 
 void pafsrand(unsigned int seed, int set) // = 0
 {
