+++ salalib/tigerp.h
@@ -26,7 +26,7 @@
 // C,D etc are not currently parsed, but given the nice file format 
 // (thank you US Census Bureau!) they can easily be added
 
-class TigerChain : public pvecline
+class TigerChain : public prefvec<Line>
 {
 public:
    TigerChain() {;}
